Dataset
A comprehensive dataset of annual time series data from 2000 to 2022 was compiled to conduct a cointegration, causality, and connectedness analysis, focusing on evaluating the impact of the 2008 GFC and the 2020 GCP on the Southeast Asia region, given its emergence as a worldwide manufacturing hub. Data on net inflows of FDI, imports of intermediate goods, technology readiness, innovation, and infrastructure were obtained from The World Bank.